+++ CHANGES.md
@@ -6,6 +6,7 @@ Glass 0.6 — ????-??-??
 
   * The `$` prefix can now be used for hexadecimal numbers.
   * Symbols can no longer start with the `$` character (bc).
+  * The `>>>` unsigned shift right operator is now supported.
 
 Glass 0.5 — 2017-01-18
 ----------------------